+++ ta4j-core/src/main/java/org/ta4j/core/analysis/CashFlow.java
@@ -133,15 +133,15 @@ public class CashFlow implements Indicator<Num> {
      */
     private void calculate(Position position, int finalIndex) {
         boolean isLongTrade = position.getEntry().isBuy();
-        int endIndex = determineEndIndex(position, finalIndex, barSeries.getEndIndex());
+        int endIndex = AnalysisUtils.determineEndIndex(position, finalIndex, barSeries.getEndIndex());
         final int entryIndex = position.getEntry().getIndex();
         int begin = entryIndex + 1;
         if (begin > values.size()) {
-            Num lastValue = values.get(values.size() - 1);
+            Num lastValue = values.getLast();
             values.addAll(Collections.nCopies(begin - values.size(), lastValue));
         }
         // Trade is not valid if net balance at the entryIndex is negative
-        if (values.get(values.size() - 1).isGreaterThan(values.get(0).getNumFactory().numOf(0))) {
+        if (values.getLast().isGreaterThan(values.getFirst().getNumFactory().numOf(0))) {
             int startingIndex = Math.max(begin, 1);
 
             int nPeriods = endIndex - entryIndex;
@@ -151,7 +151,8 @@ public class CashFlow implements Indicator<Num> {
             // Add intermediate cash flows during position
             Num netEntryPrice = position.getEntry().getNetPrice();
             for (int i = startingIndex; i < endIndex; i++) {
-                Num intermediateNetPrice = addCost(barSeries.getBar(i).getClosePrice(), avgCost, isLongTrade);
+                Num intermediateNetPrice = AnalysisUtils.addCost(barSeries.getBar(i).getClosePrice(), avgCost,
+                        isLongTrade);
                 Num ratio = getIntermediateRatio(isLongTrade, netEntryPrice, intermediateNetPrice);
                 values.add(values.get(entryIndex).multipliedBy(ratio));
             }
@@ -163,7 +164,8 @@ public class CashFlow implements Indicator<Num> {
             } else {
                 exitPrice = barSeries.getBar(endIndex).getClosePrice();
             }
-            Num ratio = getIntermediateRatio(isLongTrade, netEntryPrice, addCost(exitPrice, avgCost, isLongTrade));
+            Num ratio = getIntermediateRatio(isLongTrade, netEntryPrice,
+                    AnalysisUtils.addCost(exitPrice, avgCost, isLongTrade));
             values.add(values.get(entryIndex).multipliedBy(ratio));
         }
     }
@@ -214,51 +216,15 @@ public class CashFlow implements Indicator<Num> {
     }
 
     /**
-     * Adjusts (intermediate) price to incorporate trading costs.
-     *
-     * @param rawPrice    the gross asset price
-     * @param holdingCost share of the holding cost per period
-     * @param isLongTrade true, if the entry trade type is BUY
-     */
-    static Num addCost(Num rawPrice, Num holdingCost, boolean isLongTrade) {
-        Num netPrice;
-        if (isLongTrade) {
-            netPrice = rawPrice.minus(holdingCost);
-        } else {
-            netPrice = rawPrice.plus(holdingCost);
-        }
-        return netPrice;
-    }
-
-    /**
      * Pads {@link #values} with its last value up until {@code endIndex}.
      *
      * @param endIndex the end index
      */
     private void fillToTheEnd(int endIndex) {
         if (endIndex >= values.size()) {
-            Num lastValue = values.get(values.size() - 1);
+            Num lastValue = values.getLast();
             values.addAll(Collections.nCopies(barSeries.getEndIndex() - values.size() + 1, lastValue));
         }
     }
 
-    /**
-     * Determines the valid final index to be considered.
-     *
-     * @param position   the position
-     * @param finalIndex index up until cash flows of open positions are considered
-     * @param maxIndex   maximal valid index
-     */
-    static int determineEndIndex(Position position, int finalIndex, int maxIndex) {
-        int idx = finalIndex;
-        // After closing of position, no further accrual necessary
-        if (position.getExit() != null) {
-            idx = Math.min(position.getExit().getIndex(), finalIndex);
-        }
-        // Accrual at most until maximal index of asset data
-        if (idx > maxIndex) {
-            idx = maxIndex;
-        }
-        return idx;
-    }
 }